abort.go append.go continue.go hack.go kill.go new_pull_request.go prepend.go prune_branches.go rename_branch.go ship.go skip.go sync.go undo.go core.go handle_unfinished_state.go auto_abort.go core.go errored.go execute.go finished.go print_footer.go
