+++ src/cmd/abort.go
@@ -11,7 +11,7 @@ import (
 	"github.com/git-town/git-town/v9/src/hosting"
 	"github.com/git-town/git-town/v9/src/messages"
 	"github.com/git-town/git-town/v9/src/steps"
-	runvm "github.com/git-town/git-town/v9/src/vm/interpreter"
+	"github.com/git-town/git-town/v9/src/vm/interpreter"
 	"github.com/git-town/git-town/v9/src/vm/persistence"
 	"github.com/git-town/git-town/v9/src/vm/state"
 	"github.com/spf13/cobra"
@@ -54,7 +54,7 @@ func executeAbort(debug bool) error {
 	if err != nil {
 		return err
 	}
-	return runvm.Execute(runvm.ExecuteArgs{
+	return interpreter.Execute(interpreter.ExecuteArgs{
 		RunState:                &abortRunState,
 		Run:                     &repo.Runner,
 		Connector:               config.connector,
