+++ src/cmd/ship.go
@@ -15,7 +15,7 @@ import (
 	"github.com/git-town/git-town/v9/src/step"
 	"github.com/git-town/git-town/v9/src/steps"
 	"github.com/git-town/git-town/v9/src/validate"
-	runvm "github.com/git-town/git-town/v9/src/vm/interpreter"
+	"github.com/git-town/git-town/v9/src/vm/interpreter"
 	"github.com/git-town/git-town/v9/src/vm/state"
 	"github.com/spf13/cobra"
 )
@@ -96,7 +96,7 @@ func executeShip(args []string, message string, debug bool) error {
 		InitialActiveBranch: initialBranchesSnapshot.Active,
 		RunSteps:            shipSteps(config, message),
 	}
-	return runvm.Execute(runvm.ExecuteArgs{
+	return interpreter.Execute(interpreter.ExecuteArgs{
 		RunState:                &runState,
 		Run:                     &repo.Runner,
 		Connector:               config.connector,
