+++ src/validate/handle_unfinished_state.go
@@ -10,7 +10,7 @@ import (
 	"github.com/git-town/git-town/v9/src/hosting"
 	"github.com/git-town/git-town/v9/src/messages"
 	"github.com/git-town/git-town/v9/src/undo"
-	runvm "github.com/git-town/git-town/v9/src/vm/interpreter"
+	"github.com/git-town/git-town/v9/src/vm/interpreter"
 	"github.com/git-town/git-town/v9/src/vm/persistence"
 	"github.com/git-town/git-town/v9/src/vm/state"
 )
@@ -63,7 +63,7 @@ type UnfinishedStateArgs struct {
 
 func abortRunstate(runState *state.RunState, args UnfinishedStateArgs) (bool, error) {
 	abortRunState := runState.CreateAbortRunState()
-	return true, runvm.Execute(runvm.ExecuteArgs{
+	return true, interpreter.Execute(interpreter.ExecuteArgs{
 		Connector:               args.Connector,
 		Debug:                   args.Debug,
 		Lineage:                 args.Lineage,
@@ -85,7 +85,7 @@ func continueRunstate(runState *state.RunState, args UnfinishedStateArgs) (bool,
 	if repoStatus.Conflicts {
 		return false, fmt.Errorf(messages.ContinueUnresolvedConflicts)
 	}
-	return true, runvm.Execute(runvm.ExecuteArgs{
+	return true, interpreter.Execute(interpreter.ExecuteArgs{
 		Connector:               args.Connector,
 		Debug:                   args.Debug,
 		InitialBranchesSnapshot: args.InitialBranchesSnapshot,
@@ -106,7 +106,7 @@ func discardRunstate(rootDir domain.RepoRootDir) (bool, error) {
 
 func skipRunstate(runState *state.RunState, args UnfinishedStateArgs) (bool, error) {
 	skipRunState := runState.CreateSkipRunState()
-	return true, runvm.Execute(runvm.ExecuteArgs{
+	return true, interpreter.Execute(interpreter.ExecuteArgs{
 		Connector:               args.Connector,
 		Debug:                   args.Debug,
 		InitialBranchesSnapshot: args.InitialBranchesSnapshot,
