+++ src/vm/interpreter/core.go
@@ -1,2 +1,2 @@
 // Package runvm implements the virtual machine for executing steps of Git Town commands.
-package runvm
+package interpreter
